pyproject.toml
[tool.poetry]
name = "repl_python3_beta_ISRO-space-challenge"
version = "0.1.0"
description = ""
authors = ["C3TMLive <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
